Remove src/gui/slices.o from repo. Added PrimitiveList::add() function to add a new primitive to the list. Added Viewer::createGridSurface(). Fixed - fixedAlpha_ value was declared as an int in Collection, but expected as a double elsewhere (changed to be double throughout). Fixed - Data window was not updated when hidden.

+++ src/base/collection.h
@@ -196,7 +196,7 @@ class Collection : public ListItem<Collection>
 	// Current alpha control
 	AlphaControl alphaControl_;
 	// Fixed alpha value (for FixedAlpha option)
-	int fixedAlpha_;
+	double fixedAlpha_;
 	// Flag indicating that colourscale is valid (and doesn't need to be regenerated)
 	bool colourScaleValid_;
 
@@ -213,7 +213,7 @@ class Collection : public ListItem<Collection>
 	const ColourScalePoint* colourScalePoint(ColourSource source, int index = -1);
 	// Return colour of point specified
 	QColor colourScalePointColour(ColourSource source, int index = -1);
-	// Return value of point specified
+	// Retu5rn value of point specified
 	double colourScalePointValue(ColourSource source, int index = -1);
 	// Add empty point to end of custom colourscale
 	void addCustomColourScalePoint();
